Modularized file structure, set up rudementary add internship functionality
