+++ app/seeker/Templates/add_internship.html
@@ -16,6 +16,9 @@
                     {{ form.year(placeholder="Year", class="form-control  form-control-sm") }}
                 </div>
                 <div>
+                    {{ form.location(placeholder="Location", class="form-control  form-control-sm") }}
+                </div>
+                <div>
                     {{ form.additional_information(placeholder="Add Info", class="form-control  form-control-sm") }}
                 </div>
         </div>
