+++ app/seeker/Templates/internships.html
@@ -6,6 +6,7 @@
             {{internship.company}}
             {{internship.term}}
             {{internship.year}}
+            {{internship.location}}
             {{internship.additional_information}}
         </p>
     {% endfor %}
