+++ app/seeker/models/Internship.py
@@ -1,10 +1,4 @@
-"""
-models.py
-====================================
-The models that go into the database
-"""
-
-from . import db
+from .. import db
 
 
 class Internship(db.Model):
@@ -12,4 +6,5 @@ class Internship(db.Model):
     company = db.Column(db.String(100), index=True, nullable=False)
     term = db.Column(db.String(100), index=True, nullable=False)
     year = db.Column(db.Integer, index=True, nullable=False)
+    location = db.Column(db.String(100), index=True, nullable=False)
     additional_info = db.Column(db.String(800), index=True, nullable=False)
