+++ app/seeker/routes/internship.py
@@ -7,24 +7,18 @@ The url routes that render html templates, and handles forms and database manipu
 
 from __future__ import print_function
 from flask import current_app as app
-from flask import render_template, url_for, flash, redirect, request
-from . import db
-from .forms import InternshipForm
-from .models import Internship
+from flask import render_template, url_for, redirect
+from .. import db
+from ..forms.InternshipForm import InternshipForm
+from ..models.Internship import Internship
 
-@app.route('/')
-@app.route('/home')
-def home():
-    """
-    Route for the home page
-    """
-    return render_template('home.html')
 
 @app.route('/add_internship', methods=['GET', 'POST'])
 def add_internship():
     form = InternshipForm()
     if form.validate_on_submit():
-        internship = Internship(company=form.company.data, term=form.term.data, year=form.year.data, additional_info=form.additional_information.data)
+        internship = Internship(company=form.company.data, term=form.term.data, year=form.year.data, location=form.location.data,
+                                additional_info=form.additional_information.data)
         db.session.add(internship)
         db.session.commit()
         return redirect(url_for('home'))
@@ -33,6 +27,5 @@ def add_internship():
 
 @app.route('/internships')
 def internships():
-    internships = db.session.query(Internship)
-    print(internships)
-    return render_template('internships.html', internsips=internships)
+    internships = db.session.query(Internship).all()
+    return render_template('internships.html', internships=internships)
